Refactor residual >700-line files below the size ratchet (warren-9a61)
Split the three remaining in-scope oversized files under the
file-size budget and lower the ratchet accordingly:

- src/server/handlers/plots.workbench.test.ts (749) ->
  plots.workbench.answer.test.ts + plots.workbench.formalize.test.ts,
  sharing the extracted plots.workbench.harness.ts.
- src/projects/refresh.test.ts (709) -> refresh.test.ts +
  refresh.plot-preserve.test.ts, sharing refresh.test-helpers.ts.
- scripts/acceptance/scenarios/32-plot-workbench-loop.ts (706) ->
  scenario file + 32-plot-workbench-loop.helpers.ts.

Removed the three now-stale grandfather entries from
scripts/file-size-budgets.json (ratchet only goes down). The three
UI files are out of scope: check-file-sizes.ts excludes src/ui/ via
EXCLUDE_PATH_PREFIXES, so they remain implicitly grandfathered.
