+++ .mulch/expertise/build.jsonl
@@ -101,3 +101,4 @@
 {"type":"failure","classification":"tactical","recorded_at":"2026-05-29T16:37:26.297Z","evidence":{"seeds":"warren-436a","commit":"ed4886ac2207363cdfcc40a04d7eb9f07d5fb6a6"},"description":"In a fresh burrow workspace without node_modules, 150+ test files error with 'Cannot find package drizzle-orm/zod/pino/@os-eco/burrow-cli' and check:coverage collapses to ~85% functions vs the 87.09% floor — a false red gate.","resolution":"Run 'bun install' (root + src/ui) before trusting check:coverage / check:all. Confirms mx-6915a8.","id":"mx-fea146"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-29T16:37:36.620Z","evidence":{"seeds":"warren-436a","commit":"ed4886ac2207363cdfcc40a04d7eb9f07d5fb6a6"},"content":"Run Analytics dashboard Phase 2 (warren-436a / pl-ad0f step 10): src/ui/src/pages/run-analytics/Insights.tsx (severity-coded InsightCallouts cards) + CommandMining.tsx (CommandCategoryChart bar + StuckCommandTable leaderboard with os-eco row highlighting via Badge variant=active) consume a second GET /analytics/behavior round-trip (runAnalyticsApi.behavior). Wire shapes mirror src/runs/analytics/{command-mining,insights}.ts in src/ui/src/api/client.ts. The behavior query is independent of the runs query so the fast view renders first.","id":"mx-66be6d"}
 {"type":"reference","classification":"tactical","recorded_at":"2026-05-29T16:46:40.327Z","evidence":{"commit":"d078c8797d79f7d05dcab866833ac098f7470a2e"},"name":"run-analytics-final-qa-gate","description":"warren-9d89 / pl-ad0f step 11 (final QA): Run Analytics dashboard gate is green end-to-end — bun run check:all exits 0 with coverage functions 87.89%/lines 91.11% above floors, gen:docs/gen:openapi report 53 routes in sync, bundle-size guard ok (RunAnalytics chunk ~118KiB gzip from recharts, within budget). No source changes were needed; the only red came from an incomplete node_modules — bun install (root + src/ui) is a prerequisite for the coverage gate since missing drizzle-orm/zod/pino/js-yaml/@os-eco/burrow-cli/pg error out ~157 test files and deflate function coverage to ~85%.","id":"mx-06b98e"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-29T18:16:12.585Z","evidence":{"commit":"13fb467b5edf7588c8faf29aae724c595444b5a9"},"content":"warren-9a61 (pl-df2f follow-up to warren-1e19): split residual >700-line in-scope files below the 500-line ratchet. Pattern: extract shared test harness/helpers into a non-.test sibling (kebab/dotted name, no bun:test import) and split the suite by describe-block into per-domain .test files. Done here: src/server/handlers/plots.workbench.test.ts -> plots.workbench.{answer,formalize}.test.ts + plots.workbench.harness.ts; src/projects/refresh.test.ts -> refresh.test.ts + refresh.plot-preserve.test.ts + refresh.test-helpers.ts; scripts/acceptance/scenarios/32-plot-workbench-loop.ts -> + 32-plot-workbench-loop.helpers.ts. Removed the three stale grandfather entries from scripts/file-size-budgets.json (ratchet only goes down). The three UI files (api/types.ts, RunDetail.tsx, api/client.ts) are OUT of scope: check-file-sizes.ts excludes src/ui/ via EXCLUDE_PATH_PREFIXES, so they're implicitly grandfathered and need no budget entry.","id":"mx-93f0e4"}
