+++ scripts/file-size-budgets.json
@@ -7,14 +7,12 @@
 		"scripts/acceptance/scenarios/25-plot-roundtrip.ts": 556,
 		"scripts/acceptance/scenarios/26-plan-run-roundtrip.ts": 563,
 		"scripts/acceptance/scenarios/28-plot-list-and-create.ts": 559,
-		"scripts/acceptance/scenarios/32-plot-workbench-loop.ts": 706,
 		"src/client/client.ts": 569,
 		"src/db/repos/runs.ts": 610,
 		"src/diagnostics/checks.test.ts": 653,
 		"src/diagnostics/checks.ts": 616,
 		"src/plan-runs/coordinator.ts": 587,
 		"src/plots/aggregate.ts": 506,
-		"src/projects/refresh.test.ts": 709,
 		"src/runs/pause.test.ts": 616,
 		"src/runs/pr.test.ts": 604,
 		"src/runs/pr.ts": 659,
@@ -23,7 +21,6 @@
 		"src/server/handlers/projects.settings.test.ts": 685,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
 		"src/server/handlers/plots.list.test.ts": 700,
-		"src/server/handlers/plots.workbench.test.ts": 766,
 		"src/server/integration.multi-worker.test.ts": 553,
 		"src/server/server.test.ts": 536
 	}
